# Nest ifs for dtype and uop in pattern matcher