@Qazalin 2024-06-04
confirmed this is faster than master.